Return array of objects for fonts
So we have name and URL together.

Note that I'm assuming "italic === `ital` axis" currently. But
this isn't really always the case. For Fraunces, a regular and
an italic font are served. For Recursive, the same font file, but
with `font-style: oblique 0deg 15deg;` set. This is indicated as
a `slnt` axis, and doesn't seem to have any `ital`. The Google
Fonts website doesn't even expose italics for Recursive.

But looks like italic versions are requested with axis notation
anyway in the Google Fonts API, e.g. for Fraunces regular:

family=Fraunces:ital@1;

So, think about how to deal with this :-)