Add example HTML file to load actual fonts
This is needed because we can't fetch Google Fonts CSS from Node,
as it will return a convoluted stylesheet with hundreds of @font-face
declarations for the font. (I don't know for whom this is useful?)

Fetching the CSS from within the browser makes sure you'll get the
font specifically for that browser.

Unfortunately we can't fetch fonts.google.com/metadata/fonts from
within the browser because of CORS. Otherwise this entire thing
could be ran from within the browser.